+++ src/components/index.js
@@ -1,0 +1,19 @@
+import InfoAlert from "./Alerts/Info/InfoAlert";
+import SuccessAlert from "./Alerts/Success/SuccessAlert";
+import Filters from "./filters/Filters";
+import MiniAddressCard from "./mini address card/MiniAddressCard";
+import MiniProductCard from "./mini product card/MiniProductCard";
+import Navbar from "./navbar/Navbar";
+import Rating from "./rating/Rating";
+import Search from "./search/Search";
+
+export {
+  InfoAlert,
+  SuccessAlert,
+  Filters,
+  MiniAddressCard,
+  MiniProductCard,
+  Navbar,
+  Rating,
+  Search,
+};
