--- src/components/mockman/MockAPI.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-import Mockman from "mockman-js";
-function MockAPI() {
-  return (
-    <div className="MockAPI">
-      <Mockman />
-    </div>
-  );
-}
-
-export default MockAPI;
