+++ src/pages/wishlist/Wishlist.jsx
@@ -2,8 +2,7 @@ import {useState} from 'react'
 import classes from "./wishlist.module.css";
 import {useCart,useWishlist} from '../../context';
 import {useNavigate} from 'react-router-dom'
-import  InfoAlert from '../../components/Alerts/Info/InfoAlert'
-import SuccessAlert from '../../components/Alerts/Success/SuccessAlert';
+import  {InfoAlert,SuccessAlert} from '../../components'
 const Wishlist = () => {
   const {wishlistState,deleteFromWishlist} = useWishlist();
   const {cartState,qtyIncDec,addToCart} = useCart();
